code review and simplification based on the feedback.
