--- spring-core/src/main/java/com/baeldung/injectiontypes/javaconfig/domain/setter/Uranus.java
@@ -1,18 +1,0 @@
-package com.baeldung.injectiontypes.javaconfig.domain.setter;
-
-
-public class Uranus {
-    private Integer diameter;
-
-    public Uranus(Integer diameter) {
-        this.diameter = diameter;
-    }
-
-    public Integer getDiameter() {
-        return diameter;
-    }
-
-    public void setDiameter(Integer diameter) {
-        this.diameter = diameter;
-    }
-}
