--- spring-core/src/main/java/com/baeldung/injectiontypes/xmlconfig/domain/constructor/Moon.java
@@ -1,18 +1,0 @@
-package com.baeldung.injectiontypes.xmlconfig.domain.constructor;
-
-
-public class Moon {
-    private Integer diameter;
-
-    public Moon(Integer diameter) {
-        this.diameter = diameter;
-    }
-
-    public Integer getDiameter() {
-        return diameter;
-    }
-
-    public void setDiameter(Integer diameter) {
-        this.diameter = diameter;
-    }
-}
