--- spring-core/src/main/java/com/baeldung/injectiontypes/xmlconfig/domain/setter/Jupiter.java
@@ -1,18 +1,0 @@
-package com.baeldung.injectiontypes.xmlconfig.domain.setter;
-
-
-public class Jupiter {
-    private Integer diameter;
-
-    public Jupiter(Integer diameter) {
-        this.diameter = diameter;
-    }
-
-    public Integer getDiameter() {
-        return diameter;
-    }
-
-    public void setDiameter(Integer diameter) {
-        this.diameter = diameter;
-    }
-}
