+++ spring-core/src/main/java/com/baeldung/injectiontypes/xmlconfig/setter/SolarSystemSetterDi.java
@@ -1,0 +1,17 @@
+package com.baeldung.injectiontypes.xmlconfig.setter;
+
+
+import com.baeldung.injectiontypes.domain.Jupiter;
+
+public class SolarSystemSetterDi {
+
+    private Jupiter jupiter;
+
+    public void printDiameters() {
+        System.out.println("Diameter of the jupiter is :" + jupiter.getDiameter() + " km");
+    }
+
+    public void setJupiter(Jupiter jupiter) {
+        this.jupiter = jupiter;
+    }
+}
