--- spring-core/src/main/resources/AppConfigSetter.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
-
-    <bean id="jupiter" class="com.baeldung.injectiontypes.xmlconfig.domain.setter.Jupiter">
-        <constructor-arg value="50724"/>
-    </bean>
-
-    <bean id="uranus" class="com.baeldung.injectiontypes.xmlconfig.domain.setter.Uranus">
-        <constructor-arg value="139822"/>
-    </bean>
-
-    <bean id="solarSystemSetterDi" class="com.baeldung.injectiontypes.xmlconfig.domain.setter.SolarSystemSetterDi">
-        <property name="jupiter" ref="jupiter">
-
-        </property>
-        <property name="uranus" ref="uranus">
-
-        </property>
-    </bean>
-</beans>
